// src/services/items/item.types.ts
export enum ItemTypes {
  GET_ITEM_LIST_START = 'GET_ITEM_LIST_START',
  GET_ITEM_LIST_SUCCESS = 'GET_ITEM_LIST_SUCCESS',
  GET_ITEM_LIST_FAIL = 'GET_ITEM_LIST_FAIL',

  ADD_ITEM_START = 'ADD_ITEM_LIST',
  ADD_ITEM_SUCCESS = 'ADD_ITEM_SUCCESS',
  ADD_ITEM_FAIL = 'ADD_ITEM_FAIL',

  REMOVE_ITEM_START = 'REMOVE_ITEM_START',
  REMOVE_ITEM_SUCCESS = 'REMOVE_ITEM_SUCCESS',
  REMOVE_ITEM_FAIL = 'REMOVE_ITEM_FAIL',
}